# services/api-gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Eucastan_eucastanpay_services_api-gateway_internal_dto_request_account.DepositRequest:
    properties:
      account_id:
        type: string
      account_no:
        type: integer
      amount:
        type: integer
      currency:
        type: string
    required:
    - account_id
    - account_no
    - amount
    type: object
...
  github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    name: Stanley Emeh
  description: API Gateway for the EucastanPay Microservices Platform.
  title: EucastanPay API Gateway
  version: "1.0"
paths:
...
  /account/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Account Details with accountID and userID
      tags:
      - Account
...
  /admin/account/{id}/balance:
    get:
      consumes:
      - application/json
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Account owner's balance
      tags:
      - Account
  /admin/account/{id}/pay-in:
    put:
      consumes:
      - application/json
      parameters:
      - description: Deposit information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_dto_request_account.DepositRequest'
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Pay into owner's account
      tags:
      - Account
  /admin/account/{id}/withdraw:
    put:
      consumes:
      - application/json
      parameters:
      - description: Deposit information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_dto_request_account.DepositRequest'
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "409":
          description: Insufficient balance
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Withdrawing Cash from account
      tags:
      - Account
  /admin/accounts:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Eucastan_eucastanpay_services_api-gateway_internal_httpx.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Account Details for all users
      tags:
      - Account
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"